Fix #50 unit per cell as double
Include to calculate all distance in double (except when calculating pixels of course)

+++ src/main/java/net/rptools/maptool/client/walker/ZoneWalker.java
@@ -32,7 +32,7 @@ public interface ZoneWalker {
 
   public boolean isWaypoint(CellPoint point);
 
-  public int getDistance();
+  public double getDistance();
 
   public Path<CellPoint> getPath();
